modify rpc.call arguments
modify request arguments of rpc.call and deal the change return data from seele node
